288 edit lowq2 detector and vacuum geometries (#290)
### Briefly, what does this PR introduce?
An updated Low Q2 tagger implementation which is described using single
xml/cpp files and can be exported into CAD

### What kind of change does this PR introduce?
- [x] Bug fix (issue #288 )
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
Changed the detector ID for taggers to a single value

### Does this PR change default behavior?
