+++ compact/far_backward/taggers.xml
@@ -3,39 +3,60 @@
 
   <define>
 
-    <constant name="tracker_pixel_size" value="55*um"/>
+   <constant name="tracker_pixel_size" value="55*um"/>
 
   </define>
 
   <detectors>
 
-    <detector id="Tagger1_ID" type="BackwardsTracker" name="TagTrack1"
-      width="Tagger1_Width/2"
-      height="Tagger1_Height/2"
-      length="Tagger1_Length"
-      place_into="BackwardsTaggerVacuum"
-      place_into2="Tagger1"
-      readout="TaggerTrackerHits">
-      <trackLayer  id="0" type="timepix" z="0*mm"   sensor_thickness="400*um"/>
-      <trackLayer  id="1" type="timepix" z="300*mm" sensor_thickness="400*um"/>
-      <trackLayer  id="2" type="timepix" z="600*mm" sensor_thickness="400*um"/>
-      <trackLayer  id="3" type="timepix" z="900*mm" sensor_thickness="400*um"/>
-      <!--windowLayer id="1" type="window"  z="0*mm"   thickness="1*mm" material="Copper"/-->
-    </detector>
+    <comment> Main vacuum volume </comment>
+    <detector id="TaggerTracker_ID" name="BackwardsTaggerStation" type="BackwardsTagger" wall="2*mm" lumi="true" vis="BackwardsBox" readout="TaggerTrackerHits">
+      <focus      x="Dipole_Focus_X"    y="Dipole_Focus_Y"    z="Dipole_Focus_Z" />
+      <bounding
+        xmin="Vacuum_BB_MinX"
+        xmax="Vacuum_BB_MaxX"
+        ymin="Vacuum_BB_MinY"
+        ymax="Vacuum_BB_MaxY"
+        zmin="Vacuum_BB_MinZ"
+        zmax="Vacuum_BB_MaxZ"
+      />
+      <rotation   theta="Beam_Theta"/>
+      <dimensions xL="Beam_WidthL" xR="Beam_WidthR" y="Beam_Height" z="Beam_Length"/>
 
+      <comment> Vacuum volume joining with lumi </comment>
+      <exitdim    x="Exit_Width"   y="Exit_Height"  lumiZ="LumiWin_Z" lumiR="LumiWin_R" maxTheta="Exit_Theta"/>
 
-    <detector id="Tagger2_ID" type="BackwardsTracker" name="TagTrack2"
-      width="Tagger2_Width/2"
-      height="Tagger2_Height/2"
-      length="Tagger2_Length"
-      place_into="BackwardsTaggerVacuum"
-      place_into2="Tagger2"
-      readout="TaggerTrackerHits">
-      <trackLayer  id="0" type="timepix" z="0*mm"   sensor_thickness="400*um"/>
-      <trackLayer  id="1" type="timepix" z="300*mm" sensor_thickness="400*um"/>
-      <trackLayer  id="2" type="timepix" z="600*mm" sensor_thickness="400*um"/>
-      <trackLayer  id="3" type="timepix" z="900*mm" sensor_thickness="400*um"/>
-      <!--windowLayer id="1" type="window"  z="0*mm"   thickness="1*mm" material="Copper"/-->
+      <comment> Tagger stations and extended vacuum volume </comment>
+      <module id="1"
+        modname="Tagger1"
+        theta_min="Tagger1_Min_Theta"
+        max_align="false"
+        offset_min="Tagger1_Min_Offset"
+        extend_vacuum="true"
+        >
+        <dimensions x="Tagger1_Width/2" y="Tagger1_Height/2" z="Tagger1_Length"/>
+        <trackLayer  id="0" type="timepix" z="0*mm"   sensor_thickness="400*um"/>
+        <trackLayer  id="1" type="timepix" z="300*mm" sensor_thickness="400*um"/>
+        <trackLayer  id="2" type="timepix" z="600*mm" sensor_thickness="400*um"/>
+        <trackLayer  id="3" type="timepix" z="900*mm" sensor_thickness="400*um"/>
+        <!--windowLayer id="1" type="window"  z="0*mm"   thickness="1*mm" material="Copper"/-->
+      </module>
+
+      <module id="2"
+        modname="Tagger2"
+        theta_min="Tagger2_Min_Theta"
+        max_align="true"
+        offset_min="Tagger2_Min_Offset"
+        extend_vacuum="true"
+        >
+        <dimensions x="Tagger2_Width/2" y="Tagger2_Height/2" z="Tagger2_Length"/>
+        <trackLayer  id="0" type="timepix" z="0*mm"   sensor_thickness="400*um"/>
+        <trackLayer  id="1" type="timepix" z="300*mm" sensor_thickness="400*um"/>
+        <trackLayer  id="2" type="timepix" z="600*mm" sensor_thickness="400*um"/>
+        <trackLayer  id="3" type="timepix" z="900*mm" sensor_thickness="400*um"/>
+        <!--windowLayer id="1" type="window"  z="0*mm"   thickness="1*mm" material="Copper"/-->
+      </module>
+
     </detector>
 
   </detectors>
